+++ src/Controller/PagesController.php
@@ -49,8 +49,7 @@ class PagesController extends AppController
      * Static pages must be located in `APP/View/StaticPages/`.
      * @param string $slug Page slug
      * @return \Cake\Http\Response|null|void
-     * @uses \MeCms\Utility\StaticPage::get()
-     * @uses \MeCms\Utility\StaticPage::title()
+     * @uses \MeCms\Utility\StaticPage
      */
     public function view(string $slug)
     {
@@ -60,7 +59,7 @@ class PagesController extends AppController
         if ($static) {
             $page = new Entity([
                 'category' => new Entity(['slug' => null, 'title' => null]),
-                'title' => StaticPage::title($slug),
+                'title' => StaticPage::getTitle($slug),
                 'subtitle' => null,
             ] + compact('slug'));